+++ lib/appwrite.dart
@@ -39,6 +39,8 @@ part 'services/functions.dart';
 part 'services/graphql.dart';
 part 'services/locale.dart';
 part 'services/messaging.dart';
+part 'services/presences.dart';
+part 'services/advisor.dart';
 part 'services/storage.dart';
 part 'services/tables_db.dart';
 part 'services/teams.dart';